+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/coding/returncount/InputReturnCountLambda.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.coding;
+package com.puppycrawl.tools.checkstyle.checks.coding.returncount;
 
 import java.lang.Integer;
 import java.util.Optional;
